Fix: additional change of changing the type of BigDecimal to Double

+++ src/test/java/com/project/Project/controller/review/controller/ReviewRestControllerTest.java
@@ -7,10 +7,10 @@ import com.project.Project.controller.building.dto.AddressDto;
 import com.project.Project.controller.review.dto.ReviewRequestDto;
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.room.Room;
-import com.project.Project.service.impl.BuildingServiceImpl;
-import com.project.Project.service.impl.ReviewImageServiceImpl;
-import com.project.Project.service.impl.ReviewServiceImpl;
-import com.project.Project.service.impl.RoomServiceImpl;
+import com.project.Project.service.building.impl.BuildingServiceImpl;
+import com.project.Project.service.review.impl.ReviewImageServiceImpl;
+import com.project.Project.service.review.impl.ReviewServiceImpl;
+import com.project.Project.service.room.impl.RoomServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -23,7 +23,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
 import java.io.FileInputStream;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -74,18 +73,18 @@ public class ReviewRestControllerTest {
             .deposit(800)
             .monthlyRent(50)
             .managementFee(10)
-            .netLeasableArea(BigDecimal.valueOf(13.0))
-            .traffic(BigDecimal.valueOf(3.0))
-            .buildingComplex(BigDecimal.valueOf(4.0))
-            .surrounding(BigDecimal.valueOf(1.0))
-            .internal(BigDecimal.valueOf(5.0))
-            .livingLocation(BigDecimal.valueOf(3.0))
+            .netLeasableArea(13.0)
+            .traffic(3.0)
+            .buildingComplex(4.0)
+            .surrounding(1.0)
+            .internal(5.0)
+            .livingLocation(3.0)
             .advantageKeywordList(List.of("PARKING", "VENTILATION", "TOWN_NOISE"))
             .advantageDescription(advantageDescription)
             .disadvantageKeywordList(List.of("MART_CONVENIENCE_STORE", "DAY_LIGHTING"))
             .disadvantageDescription(disadvantageDescription)
             .reviewImageList(new ArrayList<>())
-            .residenceSatisfaction(BigDecimal.valueOf(5.0))
+            .residenceSatisfaction(5.0)
             .build();
 
     // https://jhkimmm.tistory.com/31
